Move changes to test.py
